Modernize the currently unused --checkstamps fast mode for kbuildsycoca.
- fix year 2106 bug, our grand grand children will be thankful :-)
- turns out QDataStream::Qt_3_1 has broken code for qint64 (this bug was in Qt3 already),
so switch to Qt_5_3 format
- increase ksycoca version to force a rebuild, obviously.

+++ src/sycoca/ksycoca_p.h
@@ -57,7 +57,7 @@ public:
     } databaseStatus;
     bool readError;
 
-    quint32 timeStamp;
+    qint64 timeStamp;
     enum { StrategyMmap, StrategyMemFile, StrategyFile, StrategyDummyBuffer } m_sycocaStrategy;
     QString m_databasePath;
     QStringList changeList;
